fix: configure spring aspects without warning
